fixed wagmi web3 connectors flow
