+++ package.json
@@ -29,8 +29,8 @@
     "react-scripts": "^5.0.1",
     "styled-components": "^5.3.5",
     "typescript": "^4.7.4",
-    "wagmi": "^0.11.4",
-    "web3": "^1.8.1"
+    "wagmi": "^0.11.6",
+    "web3": "^1.8.2"
   },
   "scripts": {
     "start": "react-app-rewired start",
